Fully parse LMS discovery response to extract additional fields, especially http_port
This is both a bug fix and enhancement. The bug was since we start the
discovery packet with `IPAD` in certain situations Slimserver would
respond with `EIPAD...` rather than `ENAME...`, which would break server
detection. The change in discovery response is due to Slimserver only
returning the IP address when the server admin has explicitly set the IP
address to listen on [1] (rather than the server detecting the address
to listen on).

The enhancement is now we can use the `JSON` key to extract the HTTP
port hosting the Slimserver UI (along with other metadata like server
name, version, etc.). The HTTP port is critical to get album art when
Slimserver is running on a non-standard port, for example in my case
9002 because Synology reserves ports 9000 and 9001 for their
applications [2].

Since we don't run discovery if the server is specified in the config
file, add `port` and `http_port` keys to the LMSWrapper config file. These
values are not currently visible in the Beocreate2 web UI, but may be
added in the future.

[1] https://github.com/LMS-Community/slimserver/blob/8.5.1/Slim/Networking/Discovery.pm#L117
[2] https://github.com/LMS-Community/slimserver-platforms/tree/HEAD/Docker#docker-on-synology